Handle restarts for services in `fixturenet-optimism` stack (#282)
* Check existing L1 contracts deployment

* Rename volume used for generated L2 config

* Check for existing L2 geth data directory

* Cross check existing L2 config against L1 deployment config

* Verify sequencer key in existing L2 geth data directory

* Add instructions to troubleshoot corrupt L2 geth dir

* Separate out instructions to run L2 with external L1

* Update docs

+++ app/data/config/fixturenet-optimism/run-op-geth.sh
@@ -4,34 +4,50 @@ if [ -n "$CERC_SCRIPT_DEBUG" ]; then
   set -x
 fi
 
-mkdir datadir
-
-echo "pwd" > datadir/password
-
 # TODO: Add in container build or use other tool
-echo "installing jq"
+echo "Installing jq"
 apk update && apk add jq
 
-# Get SEQUENCER KEY from keys.json
+# Get SEQUENCER key from keys.json
 SEQUENCER_KEY=$(jq -r '.Sequencer.privateKey' /l2-accounts/keys.json | tr -d '"')
-echo $SEQUENCER_KEY > datadir/block-signer-key
 
-geth account import --datadir=datadir --password=datadir/password datadir/block-signer-key
+# Initialize op-geth if datadir/geth not found
+if [ -f /op-node/jwt.txt ] && [ -d datadir/geth ]; then
+  echo "Found existing datadir, checking block signer key"
 
-while [ ! -f "/op-node/jwt.txt" ]
-do
-  echo "Config files not created. Checking after 5 seconds."
-  sleep 5
-done
+  BLOCK_SIGNER_KEY=$(cat datadir/block-signer-key)
 
-echo "Config files created by op-node, proceeding with script..."
+  if [ "$SEQUENCER_KEY" = "$BLOCK_SIGNER_KEY" ]; then
+    echo "Sequencer and block signer keys match, skipping initialization"
+  else
+    echo "Sequencer and block signer keys don't match, please clear L2 geth data volume before starting"
+    exit 1
+  fi
+else
+  echo "Initializing op-geth"
 
-cp /op-node/genesis.json ./
-geth init --datadir=datadir genesis.json
+  mkdir -p datadir
+  echo "pwd" > datadir/password
+  echo $SEQUENCER_KEY > datadir/block-signer-key
+
+  geth account import --datadir=datadir --password=datadir/password datadir/block-signer-key
+
+  while [ ! -f "/op-node/jwt.txt" ]
+  do
+    echo "Config files not created. Checking after 5 seconds."
+    sleep 5
+  done
+
+  echo "Config files created by op-node, proceeding with the initialization..."
+
+  geth init --datadir=datadir /op-node/genesis.json
+  echo "Node Initialized"
+fi
 
 SEQUENCER_ADDRESS=$(jq -r '.Sequencer.address' /l2-accounts/keys.json | tr -d '"')
 echo "SEQUENCER_ADDRESS: ${SEQUENCER_ADDRESS}"
-cp /op-node/jwt.txt ./
+
+# Run op-geth
 geth \
   --datadir ./datadir \
   --http \
@@ -52,7 +68,7 @@ geth \
   --authrpc.vhosts="*" \
   --authrpc.addr=0.0.0.0 \
   --authrpc.port=8551 \
-  --authrpc.jwtsecret=./jwt.txt \
+  --authrpc.jwtsecret=/op-node/jwt.txt \
   --rollup.disabletxpoolgossip=true \
   --password=./datadir/password \
   --allow-insecure-unlock \
